Unify D3D11/GL shaders: PBR, PCF+SSAO, god rays, camera freeze, fix half2float bug
- Port PBR deferred lighting (GGX NDF, Schlick Fresnel, Smith geometry) to D3D11 HLSL
- Unify SHADOW_COMP_PASS: PCF shadow sampling + SSAO in both APIs
- Unify LIGHT_RAY_MARCHING: dynamic steps from uniform, matching bias values
- Fix GL shadow/god ray bias (0.0005->0.000005, 0.001->0.00005)
- Add [loop] attributes to 5 dynamic HLSL loops (fixes fxc X3511 unroll errors)
- Fix critical half2float unsigned underflow in D3D11 PPM readback
- Implement camera freeze for deterministic frame dumps (disable m_externalControl)
- Add COLOR4 linear depth GBuffer attachment for GL depth readback
- Remove debug logging from SaveD3D11TextureToPPM
- Results: GBuffer_Depth 0% diff, Color0 center-pixel identical, Bloom 0.6% diff

+++ T850/Framework/src/video/windows/D3DXDriver.cpp
@@ -689,7 +689,7 @@ namespace t800 {
       unsigned int mant = h & 0x3FF;
       if (exp == 0) return 0.0f;
       if (exp == 31) return sign ? -1e30f : 1e30f;
-      float f = powf(2.0f, (float)(exp - 15)) * (1.0f + mant / 1024.0f);
+      float f = powf(2.0f, (float)((int)exp - 15)) * (1.0f + mant / 1024.0f);
       return sign ? -f : f;
     };
